fix(autonomous): make the child tenant repair repeatable for rolling deployments (#7396)
A one-shot versioned migration leaves a rolling deployment exposed: a
not-yet-upgraded node can write a default-stamped child row AFTER the
repair ran on the first upgraded node, and Flyway would never rerun, so
the row would stay misattributed forever - hidden from run-scoped reads
and able to wedge its run's event sequence on the global unique index,
with no self-serve recovery (the restart purge is itself tenant-filtered
and cannot delete the hidden row).

Replace V6_20260815100000000 with the repeatable
R__Realign_autonomous_child_tenant_attribution whose checksum is a
per-boot snapshot: Flyway re-applies the (idempotent, lock-light) repair
on every startup, so whatever an old node wrote during the rolling
window is realigned at the latest on the next boot of any node -
typically by the remaining boots of the same deployment. The write
window itself stays deliberately open: a database trigger would add a
second, DB-side attribution authority against the v2 application-layer
attribution principle (B3), and a repair-then-activate release train
would delay closing the live cross-tenant read gap for every install.

The migration test now targets the repeatable class, and a pure-JVM
contract test pins the Flyway semantics (versionless + non-null,
boot-stable checksum; instantiation also proves the R__ name parses).
